\name{CSA_PARAM_SPEC}
\alias{CSA_PARAM_SPEC}
\docType{data}
\title{
CSA PARAM SPEC
}
\description{
default values for PARAM SPEC
}
\usage{data("CSA_PARAM_SPEC")}
\format{
  A data frame on the following 2 variables.
  \describe{
    \item{\code{Parameter ID}}{a character vector}
    \item{\code{User provided input}}{a numerical vector}
  }
}
\examples{
data(CSA_PARAM_SPEC)
}
\keyword{datasets}
