\name{CSA_reference_xlsxAnalyzer}
\alias{CSA_reference_xlsxAnalyzer}
\title{
CSA reference xlsxAnalyzer
}
\description{
CSA reference xlsxAnalyzer
}
\usage{
CSA_reference_xlsxAnalyzer(ref_xlsx_file, input_path_hrms = NULL, PARAM = NULL,
PARAM_ID = "", checkpoint_parameter = TRUE)
}
\arguments{
  \item{ref_xlsx_file}{
ref_xlsx_file
}
  \item{input_path_hrms}{
input_path_hrms
}
  \item{PARAM}{
PARAM
}
  \item{PARAM_ID}{
PARAM_ID
}
  \item{checkpoint_parameter}{
checkpoint_parameter
}
}
\value{
  \item{ref_table}{
ref_table
}
  \item{PARAM}{
PARAM
}
  \item{checkpoint_parameter}{
checkpoint_parameter
}
}