\name{IDSL.CSA_workflow}
\alias{IDSL.CSA_workflow}
\title{
IDSL.CSA workflow
}
\description{
This function executes the CSA workflow.
}
\usage{
IDSL.CSA_workflow(spreadsheet)
}
\arguments{
  \item{spreadsheet}{
CSA spreadsheet
}
}
\value{
This function organizes the IDSL.CSA file processing for better performance using the template spreadsheet.
}