\name{negativeAdducts}
\alias{negativeAdducts}
\docType{data}
\title{
negative adducts
}
\description{
This data consists of adducts and mass differences for common ionization pathways in negative modes.
}
\usage{data("negativeAdducts")}
\format{
  A data frame on the following 2 variables.
  \describe{
    \item{\code{Adduct}}{a character vector}
    \item{\code{massAdduct}}{a numerical vector}
  }
}
\examples{
data(negativeAdducts)
}
\keyword{datasets}
