\name{positiveAdducts}
\alias{positiveAdducts}
\docType{data}
\title{
positive adducts
}
\description{
This data consists of adducts and mass differences for common ionization pathways in positive modes.
}
\usage{data("positiveAdducts")}
\format{
  A data frame on the following 2 variables.
  \describe{
    \item{\code{Adduct}}{a character vector}
    \item{\code{massAdduct}}{a numerical vector}
  }
}
\examples{
data(positiveAdducts)
}
\keyword{datasets}
